Patch by: daniel.libicki Reviewed by: zundel
this patch fixes a bug where JavaScript dates were not being correctly converted to Java dates in the Properties class, also a bug where Selection methods supporting AbstractVisualization would throw a JavaScript exception.  the patch also fills in a lot of javadoc.

+++ trunk/visualization/visualization/src/com/google/gwt/visualization/client/Properties.java
@@ -31,7 +31,7 @@ public class Properties extends JavaScriptObject {
   }-*/;
 
   public final Date getDate(String name) {
-    return new Date((int) getDateDouble(name));
+    return new Date((long) getDateDouble(name));
   }
 
   public final native double getDouble(String name) /*-{
